Right, so Steam installs currently need to be using XIVLauncher to get in.  So, due to the way XIVLauncher works, we need to preserve the proton prefix's version.  That means running proton's wine.  This is pretty beta but I don't really have a good way to test it.
On the other hand, if you used the script as it was the old way, you were potentially kind of gently putting the prefix into a shredding machine in Proton / XIVLauncher's eyes.  Maybe I can throw together some kind of test prefix born of lies tomorrow but in the mean time, *trying* to preserve steam users.